Moved `AttemptStrategy` (and related classes) to its own file. Made the `GQL` instance in `TwitchChannelPointsMiner` configurable. Added configuration examples/descriptions to the example and readme.
