+++ TwitchChannelPointsMiner/classes/gql/Integration.py
@@ -1,11 +1,11 @@
 import copy
 import logging
-import time
 import traceback
 from secrets import token_hex
-from typing import Callable, Any
+from typing import Callable, Any, Protocol
 
 import requests
+from requests import Response
 
 from TwitchChannelPointsMiner.classes.ClientSession import ClientSession
 from TwitchChannelPointsMiner.classes.Settings import FollowersOrder
@@ -41,6 +41,11 @@ from TwitchChannelPointsMiner.classes.gql.data.response.VideoPlayerStreamInfoOve
 )
 from TwitchChannelPointsMiner.constants import GQLOperations, CLIENT_ID
 from TwitchChannelPointsMiner.utils import create_chunks
+from TwitchChannelPointsMiner.utils.AttemptStrategy import (
+    SuccessResult,
+    ErrorResult,
+    AttemptStrategy,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -91,121 +96,12 @@ def parse_list[T](parse: Callable[[Any], T], value: Any) -> list[T]:
     raise InvalidJsonShapeException([], "list expected")
 
 
-class ExceptionContext:
-    """Utility for printing a stack trace outside the context of an exception handler."""
+class PostRequest(Protocol):
+    """For creating Type Hints for a function that posts GQL requests."""
 
-    def __init__(self, exception: Exception, stack_trace: str | None):
-        self.exception = exception
-        self.stack_trace = stack_trace
-
-    def __repr__(self):
-        if self.stack_trace is not None:
-            return f"{self.stack_trace}\n{self.exception}"
-        else:
-            return f"{self.exception}"
-
-    def __eq__(self, other):
-        return (
-            isinstance(other, ExceptionContext)
-            and self.exception == other.exception
-            and self.stack_trace == other.stack_trace
-        )
-
-
-class SuccessResult[TResult]:
-    """Returned when the result of `make_attempts` was successful."""
-
-    def __init__(self, errors: list[ExceptionContext], result: TResult):
-        self.errors = errors
-        """Any errors that occurred."""
-        self.result = result
-        """The result."""
-
-    @property
-    def attempts(self):
-        """The number of attempts made."""
-        return len(self.errors) + 1
-
-    def __repr__(self):
-        return f"SuccessResult({self.__dict__})"
-
-    def __eq__(self, other):
-        if isinstance(other, SuccessResult) and len(self.errors) == len(other.errors):
-            for index in range(len(self.errors)):
-                if self.errors[index] != other.errors[index]:
-                    return False
-            return self.result == other.result
-        return False
-
-
-class ErrorResult:
-    """Returned when the result of `make_attempts` was 1 or more errors."""
-
-    def __init__(self, errors: list[ExceptionContext]):
-        self.errors = errors
-        """The errors in the order they occurred."""
-
-    @property
-    def attempts(self):
-        """The number of attempts made."""
-        return len(self.errors)
-
-    def __repr__(self):
-        return f"ErrorResult({self.__dict__})"
-
-    def __eq__(self, other):
-        if isinstance(other, ErrorResult) and len(self.errors) == len(other.errors):
-            for index in range(len(self.errors)):
-                if self.errors[index] != other.errors[index]:
-                    return False
-            return True
-        return False
-
-
-class AttemptStrategy:
-    """Handles making an attempt at something multiple times by catching Exceptions and validating the Result."""
-
-    def __init__(self, attempts: int = 3, attempt_interval_seconds: int = 1):
-        self.attempts = attempts
-        """The number of attempts that should be made."""
-        self.attempt_interval_seconds = attempt_interval_seconds
-        """The number of seconds to wait between attempts."""
-
-    def make_attempts[TResult](
-        self,
-        attempt: Callable[[], TResult],
-        validate: Callable[[TResult], None],
-        retryable: Callable[[Exception], bool],
-        exception_context: Callable[[Exception], str | None],
-    ) -> SuccessResult | ErrorResult:
-        """
-        Calls `attempt` up to `self.attempts` times until either a successful attempt is made or the maximum number of
-        attempts have been made.
-        :param attempt: The functon to attempt.
-        :param validate: Function to check if the result is valid. Should throw an Exception if not.
-        :param retryable: Function that returns True if a given Error can be retried.
-        :param exception_context: Function that returns a context string (or None) for a given Exception.
-        :return:
-        """
-        attempts = 0
-        errors: list[ExceptionContext] = []
-        while attempts < self.attempts:
-            attempts += 1
-            try:
-                result = attempt()
-                validate(result)
-                return SuccessResult(errors, result)
-            except Exception as e:
-                errors.append(ExceptionContext(e, exception_context(e)))
-                if not retryable(e):
-                    logger.debug(f"Error cannot be retried: {e}")
-                    break
-            if attempts >= self.attempts:
-                # Break early to avoid sleeping
-                break
-            else:
-                time.sleep(self.attempt_interval_seconds)
-        return ErrorResult(errors)
+    def __call__(
+        self, url: str, json: dict | list, headers: dict[str, str]
+    ) -> Response: ...
 
 
 class GQL:
@@ -217,8 +113,8 @@ class GQL:
         self,
         client_session: ClientSession,
         attempt_strategy: AttemptStrategy | None = None,
-        parser: Parser = Parser(),
-        post_request=requests.post,
+        parser: Parser | None = None,
+        post_request: PostRequest | None = None,
     ):
         self.client_session = client_session
         """The client session for making requests."""
@@ -228,9 +124,9 @@ class GQL:
             else AttemptStrategy(attempts=3, attempt_interval_seconds=1)
         )
         """Strategy for handling failed requests."""
-        self.parser = parser
+        self.parser = Parser() if parser is None else parser
         """The parser for parsing GQL responses."""
-        self.post_request = post_request
+        self.post_request = requests.post if post_request is None else post_request
         """Function for posting GQL requests."""
 
     def __post_gql_request[T](
@@ -623,4 +519,28 @@ class GQL:
             GQLOperations.ContributeCommunityPointsCommunityGoal["operationName"],
             json_data,
             self.parser.parse_contribute_community_points_community_goal,
+        )
+
+
+class GQLFactory:
+    """Factory class for creating GQL objects."""
+
+    def __init__(
+        self,
+        attempt_strategy: AttemptStrategy | None = None,
+        parser: Parser | None = None,
+        post_request: PostRequest | None = None,
+    ):
+        self.attempt_strategy = attempt_strategy
+        self.parser = parser
+        self.post_request = post_request
+
+    def create(self, client_session: ClientSession) -> GQL:
+        """
+        Creates a new GQL instance.
+        :param client_session: The ClientSession for the instance.
+        :return: The instance.
+        """
+        return GQL(
+            client_session, self.attempt_strategy, self.parser, self.post_request
         )
