+++ TwitchChannelPointsMiner/utils/__init__.py
@@ -1,1 +1,2 @@
 from .Utils import *
+from .AttemptStrategy import *
